+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ However, there will be an initial period of stabilisation where this is not adhe
 
 ### New Features
 
+- Added a troubleshooting and FAQ page.
 - Included support for non-interactive commands, including stopping any long-running commands.
 - Added `--prebuilt-images` option to `qcb up` to bring QCrBox up using tested images from the QCrBox container
   repository.
@@ -49,6 +50,8 @@ However, there will be an initial period of stabilisation where this is not adhe
 - The application types in `qcb init` have been given more descriptive names: "CLI" -> "Non-interactive Command" and
   "GUI (Linux)" -> "Interactive GUI (Linux)". We have also reintroduced the "Interactive GUI (Windows)" application
   which lets you build an interactive application based on a Windows application running via Wine.
+- Improved error message return from command invocation/ending API endpoints when NATS is the reason for failure.
+- `PythonCallable` commands can now also return a `pathlib.Path` object in addition to a `str` of a file path.
 - And lots more!
 
 ### Issues Fixed
@@ -64,6 +67,10 @@ However, there will be an initial period of stabilisation where this is not adhe
 - Fixed an issue where the `base-ancestor` image wouldn't build because `conda` couldn't resolve the environment
   dependencies.
 - Fixed an issue where environment variables were unset causing applications not to launch
+- Fixed an issue where the output cif parameter was not being used correctly
+- Fixed an issue where `qcb init` would not create a valid template for interactive applications
+- Fixed an issue where cif files would become lost during command execution, which resulted in incorrect unified cifs
+  being returned to the registry.
 - And lots more!
 
 ### Documentation